Refactor debounce export
